@@ -6,65 +6,77 @@
     "start": "set NODE_ENV=production && node ./bin/www",
     "dev-start": "set NODE_ENV=develop && set SERVER=stage && node ./bin/www",
     "dev-build": "set NODE_ENV=develop && npm run webpack",
-    "build": "set NODE_ENV=production && npm run webpack && cd ./server && tsc && cd ..",
+    "build": "set NODE_ENV=production && npm run webpack && yarn run build-server",
+    "set-env": "set NODE_OPTIONS=--openssl-legacy-provider",
     "build-server": "cd ./server && tsc && cd ..",
     "build-watch": "set NODE_ENV=production && npm run webpack-watch",
     "test": "mocha -r ts-node/register --extension js --extension ts --timeout 10000",
     "webpack": "node ./node_modules/webpack/bin/webpack.js --colors --progress",
-    "webpack-watch": "node ./node_modules/webpack/bin/webpack.js --colors --progress --watch"
+    "webpack-watch": "node ./node_modules/webpack/bin/webpack.js --colors --progress --watch",
+    "storybook": "start-storybook -p 6006",
+    "build-storybook": "build-storybook"
   },
   "dependencies": {
-    "@babel/plugin-proposal-object-rest-spread": "^7.7.7",
-    "@babel/plugin-syntax-dynamic-import": "^7.7.4",
-    "@babel/preset-env": "^7.7.7",
-    "@babel/preset-react": "^7.7.4",
-    "@babel/preset-typescript": "^7.7.7",
-    "@emotion/react": "11.4.0",
-    "@emotion/styled": "11.3.0",
-    "@material-ui/core": "5.0.0-alpha.37",
-    "@material-ui/icons": "5.0.0-alpha.37",
-    "@material-ui/lab": "5.0.0-alpha.37",
+    "@babel/plugin-proposal-object-rest-spread": "^7.18.9",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+    "@babel/preset-env": "^7.18.9",
+    "@babel/preset-react": "^7.18.6",
+    "@babel/preset-typescript": "^7.18.6",
+    "@mui/icons-material": "^5.8.4",
+    "@mui/lab": "^5.0.0-alpha.93",
+    "@mui/material": "^5.9.2",
+    "@mui/styled-engine": "npm:@mui/styled-engine-sc@latest",
+    "@mui/styled-engine-sc": "^5.8.0",
     "@reduxjs/toolkit": "^1.8.3",
-    "@types/prop-types": "^15.7.3",
-    "@types/react": "17.0.1",
-    "@types/react-dom": "17.0.1",
+    "@types/prop-types": "^15.7.5",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
     "@types/styled-components": "^5.1.10",
     "architect": "^0.1.13",
     "axios": "^0.19.2",
+    "clsx": "^1.2.1",
     "convict": "^6.1.0",
     "convict-format-with-validator": "^6.0.1",
     "cookie-parser": "~1.4.4",
+    "core-js": "^3.24.0",
     "debug": "~2.6.9",
-    "express": "~4.16.1",
+    "express": "^4.18.1",
     "express-mysql-session": "^2.1.3",
     "express-session": "^1.17.0",
     "express-session-sqlite": "^2.1.1",
     "http-errors": "~1.6.3",
     "image-js": "^0.21.8",
     "lodash": "^4.17.15",
-    "material-ui-dropzone": "^2.4.8",
     "morgan": "~1.9.1",
     "mysql": "^2.18.1",
     "niceware": "^2.0.0",
     "prop-types": "^15.7.2",
     "pug": "^2.0.4",
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
-    "react-hook-form": "^7.8.8",
-    "react-redux": "^7.2.0",
-    "react-router-dom": "^5.2.0",
-    "redux": "^4.0.5",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.34.0",
+    "react-redux": "^8.0.2",
+    "react-router-dom": "^6.3.0",
+    "redux": "^4.2.0",
     "serve-favicon": "^2.5.0",
     "sqlite": "^4.1.1",
     "sqlite3": "^5.0.8",
     "strftime": "^0.10.0",
-    "styled-components": "^5.3.0",
+    "styled-components": "^5.3.5",
     "uuid": "^7.0.1",
     "winston": "^3.2.1"
   },
   "devDependencies": {
     "@babel/core": "^7.7.7",
     "@babel/plugin-proposal-class-properties": "^7.8.3",
+    "@storybook/addon-actions": "^6.5.10",
+    "@storybook/addon-essentials": "^6.5.10",
+    "@storybook/addon-interactions": "^6.5.10",
+    "@storybook/addon-links": "^6.5.10",
+    "@storybook/builder-webpack4": "^6.5.10",
+    "@storybook/manager-webpack4": "^6.5.10",
+    "@storybook/react": "^6.5.10",
+    "@storybook/testing-library": "^0.0.13",
     "@tsconfig/node16": "^1.0.1",
     "@types/mocha": "^7.0.1",
     "@types/node": "^13.9.1",
