@@ -1,19 +1,19 @@
 
 import { configureStore } from "@reduxjs/toolkit";
 import * as React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import {Provider} from "react-redux";
-import { Application as App } from "./Components/App"
-import reducer from "./reducer";
+import { Application as App } from "./App"
+import reducer from "./reducer/reducer";
 
-const store = configureStore({
-  reducer
-})
+const store = configureStore({reducer}); // configureStore({reducer})
 
 const rootElement = document.getElementById('root');
-ReactDOM.render(
-    <Provider store={store}>
-        <App/>
-    </Provider>, rootElement);
+const root = createRoot(rootElement);
+root.render(
+  <Provider store={store}>
+      <App/>
+  </Provider>
+);
 
 
