@@ -3,15 +3,19 @@ module.exports = function (api) {
 
   return {
     "presets": [
-      ["@babel/env", { "targets": "> 0.25%", "useBuiltIns": "usage" }],
       "@babel/preset-react",
       ["@babel/preset-typescript", { "allExtensions": true, "isTSX": true }],
-
+      ["@babel/preset-env",
+      {
+        "targets": "> 0.25%", "useBuiltIns": "usage", "corejs": "3.24.0",
+      }],
     ],
     "plugins": [
       "@babel/plugin-syntax-dynamic-import",
-      "@babel/proposal-class-properties",
-      "@babel/proposal-object-rest-spread"
+      "@babel/proposal-object-rest-spread",
+      ['@babel/plugin-proposal-class-properties', {'loose': true}],
+      ['@babel/plugin-proposal-private-methods', {'loose': true}],
+      ['@babel/plugin-proposal-private-property-in-object', {'loose': true}],
     ]
   }
 };
